+++ .gitignore
@@ -144,9 +144,8 @@ test_auth_kf.py
 .history
 *.csv
 *.pkl
+*.sqlite
 docker_build_dev_image
-tests/*
-test_*.py
 
 /pipeline_secrets
 /us_docker_build_dev_image
